- Fixed profile page and created new user profile activity - Updated library page - Added signout functionality
