+++ app/src/main/res/layout/fragment_library.xml
@@ -116,21 +116,15 @@
         android:gravity="center_horizontal"
         android:orientation="horizontal">
 
-        <Button
-            android:id="@+id/library_button"
-            android:layout_width="133dp"
-            android:layout_height="wrap_content"
-            android:layout_margin="10dp"
-            android:backgroundTint="#FFF4E5"
-            android:text="Library"
-            android:textColor="#06000A" />
 
         <Button
             android:id="@+id/profile_button"
             android:layout_width="133dp"
             android:layout_height="wrap_content"
+            android:layout_gravity="right"
             android:layout_margin="10dp"
-            android:backgroundTint="#FFF4E5"
+            android:background="@drawable/button_background"
+            android:backgroundTint="#e3d4d6"
             android:text="Profile"
             android:textColor="#06000A" />
